refactor(EventName): typescript is now aware of event names
./src/data/events/index.ts now exports the event names, in order,
as a const assertion.

We then use that in ./src/data/types.ts to generate the EventName
type.

We then update all instances of `event: string` and `slug: string`
to use the narrower EventName instead.

This commit also adds type checking to the shape of objects in the
colors array inside src/components/EventTheme/colors.tsx

This PR leaves the Tito-Widget event name as a string since it
appears to support multiple event name encodings.